+++ build-all-branches.bash
@@ -2,7 +2,10 @@
 
 # Builds and tests with all major versions we support
 
-./gradlew -Pintellij=2018.2.3 clean build
-./gradlew -Pintellij=2018.1.3 clean build
-./gradlew -Pintellij=2017.1.3 clean build
+for ij in "2018.2.3" "2018.1.3" "2017.1.3"; do
+    ./gradlew -Pintellij="$ij" clean build
+    [[ "$?" != "0" ]] && echo "Build with IntelliJ $ij failed" && exit 1
+done
+
 ./gradlew clean build
+[[ "$?" != "0" ]] && echo "Build with IntelliJ 2016.1 failed" && exit 1