+++ Changelog.md
@@ -1,6 +1,7 @@
 # Version 0.2
 - Highlighting errors in macro names and macro arguments
   (DUP, DIP, PAIR, UNPAIR, CADR, SET_CADR, MAP_CADR and variants)
+- Basic formatting of Michelson files
 
 # Version 0.1
 - Initial release